marked notices and devided actions
