+++ src/components/filters/filters.jsx
@@ -1,12 +1,8 @@
 import {connect} from "react-redux";
-import {ActionCreator} from "../../reducer.js";
-import {DEFAULT_GENRE} from '../../utils/consts';
+import {mapStateToProps, mapDispatchToProps} from './filters.connect';
 
 const Filters = (props) => {
-  const {films, onFilterClick, chosenGenre} = props;
-
-  const uniqueGenres = Array.from(new Set(films.map((film) => film.genre)));
-  const listGenres = [].concat(DEFAULT_GENRE).concat(uniqueGenres);
+  const {onFilterClick, chosenGenre, listGenres} = props;
 
   return (
     <ul className="catalog__genres-list">
@@ -35,21 +31,9 @@ Filters.propTypes = {
       })
   ).isRequired,
   onFilterClick: PropTypes.func.isRequired,
-  chosenGenre: PropTypes.string.isRequired
+  chosenGenre: PropTypes.string.isRequired,
+  listGenres: PropTypes.arrayOf(PropTypes.string).isRequired
 };
-
-const mapStateToProps = (state) => {
-  return {
-    films: state.films,
-    chosenGenre: state.chosenGenre
-  };
-};
-
-const mapDispatchToProps = (dispatch) => ({
-  onFilterClick(genre) {
-    dispatch(ActionCreator.changeFilteredGenre(genre));
-  }
-});
 
 export {Filters};
 export default connect(mapStateToProps, mapDispatchToProps)(Filters);
