+++ src/components/show-more-button/show-more-button.connect.js
@@ -1,0 +1,26 @@
+import {ActionCreator} from "../../actions.js";
+import {DEFAULT_GENRE} from '../../utils/consts';
+
+const getFilteredFilms = (films, activeGenre) => {
+  if (activeGenre === DEFAULT_GENRE) {
+    return films;
+  }
+
+  return films.filter((film) => film.genre === activeGenre);
+};
+
+const mapStateToProps = (state) => {
+  return {
+    films: getFilteredFilms(state.films, state.chosenGenre),
+    showedFilms: state.showedFilms,
+    filteredFilms: state.filteredFilms
+  };
+};
+
+const mapDispatchToProps = (dispatch) => ({
+  onShowMoreButtonClick() {
+    dispatch(ActionCreator.showeMoreFilms());
+  }
+});
+
+export {mapStateToProps, mapDispatchToProps};
