+++ src/components/show-more-button/show-more-button.jsx
@@ -1,6 +1,5 @@
 import {connect} from "react-redux";
-import {ActionCreator} from "../../reducer.js";
-import {DEFAULT_GENRE} from '../../utils/consts';
+import {mapStateToProps, mapDispatchToProps} from './show-more-button.connect';
 
 const ShowMoreButton = (props) => {
   const {onShowMoreButtonClick, showedFilms, films} = props;
@@ -37,28 +36,6 @@ ShowMoreButton.propTypes = {
       })
   ).isRequired
 };
-
-const getFilteredFilms = (films, activeGenre) => {
-  if (activeGenre === DEFAULT_GENRE) {
-    return films;
-  }
-
-  return films.filter((film) => film.genre === activeGenre);
-};
-
-const mapStateToProps = (state) => {
-  return {
-    films: getFilteredFilms(state.films, state.chosenGenre),
-    showedFilms: state.showedFilms,
-    filteredFilms: state.filteredFilms
-  };
-};
-
-const mapDispatchToProps = (dispatch) => ({
-  onShowMoreButtonClick() {
-    dispatch(ActionCreator.showeMoreFilms());
-  }
-});
 
 export {ShowMoreButton};
 export default connect(mapStateToProps, mapDispatchToProps)(ShowMoreButton);
