+++ src/helpers/helpers.js
@@ -1,0 +1,3 @@
+export const extend = (a, b) => {
+  return Object.assign({}, a, b);
+};
