+++ src/reducer.js
@@ -1,9 +1,7 @@
 import {films} from './mocks/films';
 import {DEFAULT_GENRE, DEFAULT_SHOWED_FILMS} from './utils/consts';
-
-const extend = (a, b) => {
-  return Object.assign({}, a, b);
-};
+import {extend} from './helpers/helpers';
+import {ActionType} from './actions';
 
 const initialState = {
   chosenGenre: DEFAULT_GENRE,
@@ -16,33 +14,6 @@ const initialState = {
     preview: ``
   },
   films,
-};
-
-const ActionType = {
-  CHANGE_FILTERED_GENRE: `CHANGE_FILTERED_GENRE`,
-  SET_ACTIVE_CARD: `SET_ACTIVE_CARD`,
-  SHOW_MORE_FILMS: `SHOW_MORE_FILMS`
-};
-
-const ActionCreator = {
-  setActiveFIlm: (film) => {
-    return {
-      type: ActionType.SET_ACTIVE_CARD,
-      payload: film
-    };
-  },
-  changeFilteredGenre: (genre) => {
-    return {
-      type: ActionType.CHANGE_FILTERED_GENRE,
-      payload: genre
-    };
-  },
-  showeMoreFilms: () => {
-    return {
-      type: ActionType.SHOW_MORE_FILMS,
-      payload: DEFAULT_SHOWED_FILMS
-    };
-  }
 };
 
 const reducer = (state = initialState, action) => {
@@ -75,4 +46,4 @@ const reducer = (state = initialState, action) => {
   return state;
 };
 
-export {reducer, ActionType, ActionCreator};
+export {reducer};
