+++ src/utils/consts.js
@@ -1,2 +1,3 @@
 export const DEFAULT_GENRE = `All genres`;
 export const DEFAULT_SHOWED_FILMS = 8;
+export const DEFAULT_GENRES_AMOUNT = 9;
